Derive openjdk sha and tag more reliably
In some scenarios, for example when working with a shallow clone,
the git commands used to derive the latest SHA and build tag do
not work, resulting in uninvormative output from java -version.

This commit updates the git commands used to derive the SHA and
then to map to the build tag so that they work correctly even with
a shallow clone of the repository.

Fixes: #4
Signed-off-by: Mark Stoodley <mstoodle@ca.ibm.com>