Enable cross-compile for armhf
Multiple changes were required to enable cross-compilation and composition
of the 32bit build into the 'default' lib subdirectory. This subdirectory
name is now parameterised throughout the composition steps and can be set
from custom-hook.m4 using OPENJ9_LIBS_SUBDIR.

The build jdk is not currently supported (if attempted it builds for arm
and then fails to run when used), so --with-build-jdk is used to specify
a pre-existing build jdk. Because the build jdk isn't being built it was
also necessary to prevent it being staged, otherwise the pre-existing
build jdk gets overwritten with arm binaries.

A new buildspec was added; linux_arm_linaro.

Signed-off-by: James Kingdon <jkingdon@ca.ibm.com>