Don't attempt to stage build jvm in cross compile
Updates for OpenJ9.gmk and Main.gmk to prevent attempts to stage a
build jdk during cross compiles.
This is a temporary fix - I think the right answer is probably to
detect that --with-build-jdk is being used in order to skip these
stages. And beyond that, we need to make the creation of a build
jdk work with cross compiles so that --with-build-jdk isn't needed.

Signed-off-by: James Kingdon <jkingdon@ca.ibm.com>